+++ template-cobraengineACSEgame/index.js
@@ -859,7 +859,7 @@ function applyGame(context, gameSpec) {
 function main(context) {
   applyGame(context, spec);
   context.once(() => { // put code here that should be run (once) when Vortex starts up
-
+    const api = context.api;
   });
   return true;
 }