+++ template-ue4-5game/index.js
@@ -2257,7 +2257,7 @@ function main(context) {
         if (previousLO === undefined) previousLO = loadOrder;
         if (loadOrder === previousLO) return;
         //context.api.store.dispatch(actions.setDeploymentNecessary(spec.game.id, true));
-        requestDeployment(context, spec);
+        requestDeployment(context.api, spec);
         previousLO = loadOrder;
       },
       createInfoPanel: () =>
@@ -2270,15 +2270,16 @@ function main(context) {
     });
   }
   context.once(() => { // put code here that should be run (once) when Vortex starts up
-    context.api.onAsync('did-deploy', (profileId) => didDeploy(context.api, profileId)); //*/
-    //context.api.onAsync('did-purge', (profileId) => didPurge(context.api, profileId)); //*/
+    const api = context.api;
+    api.onAsync('did-deploy', (profileId) => didDeploy(api, profileId)); //*/
+    //api.onAsync('did-purge', (profileId) => didPurge(api, profileId)); //*/
   });
   return true;
 }
 
-const requestDeployment = (context, spec) => {
-  context.api.store.dispatch(actions.setDeploymentNecessary(spec.game.id, true));
-  context.api.sendNotification({
+const requestDeployment = (api, spec) => {
+  api.store.dispatch(actions.setDeploymentNecessary(spec.game.id, true));
+  api.sendNotification({
     id: `${spec.game.id}-loadorderdeploy-notif`,
     type: 'warning',
     message: 'Deployment Required to Apply Load Order Changes',
@@ -2287,7 +2288,7 @@ const requestDeployment = (context, spec) => {
       {
         title: 'Deploy',
         action: (dismiss) => {
-          deploy(context.api)
+          deploy(api);
           dismiss();
         }
       }