+++ template-unitymelonloaderbepinex-hybrid/index.js
@@ -2151,7 +2151,8 @@ function applyGame(context, gameSpec) {
 function main(context) {
   applyGame(context, spec);
   context.once(() => { // put code here that should be run (once) when Vortex starts up
-    context.api.onAsync('did-deploy', (profileId, deployment) => { 
+    const api = context.api;
+    context.api.onAsync('did-deploy', async (profileId, deployment) => { 
       const LAST_ACTIVE_PROFILE = selectors.lastActiveProfileForGame(context.api.getState(), GAME_ID);
       if (profileId !== LAST_ACTIVE_PROFILE) return;
       bepinexInstalled = isBepinexInstalled(context.api, spec);
@@ -2176,7 +2177,7 @@ function main(context) {
       }
       return Promise.resolve();
     });
-    context.api.onAsync('did-purge', (profileId) => { 
+    context.api.onAsync('did-purge', async (profileId) => { 
       const LAST_ACTIVE_PROFILE = selectors.lastActiveProfileForGame(context.api.getState(), GAME_ID);
       if (profileId !== LAST_ACTIVE_PROFILE) return;
       if (hasCustomLoader) {