+++ game-menace/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None
 
+## [0.2.0] - 2026-02-12
+
+- Added: support for ModPackLoader mods not packaged with loader dll
+- Added: Auto-copy of ModPackLoader dlls to game folder
+
 ## [0.1.0] - 2026-02-10
 
 - Initial release